add PointerFree to the Event struct and account for tiny blocks
This change adds an overlooked field to the Event struct which we have
information about but did not surface, which is whether the object
contains pointers. It also accounts for tiny blocks, since this version
of the trace reports the tiny allocation that created a block, rather
than the block itself. Luckily, these can be easily detected
(pointer-free and smaller than 16 bytes) so we replace them with the
blocks themselves (defined as 16 byte non-array objects).