extend tiny size class sizes for immix and simulate tracking size info
This change makes it so the tiny size class for immix encompasses all
objects up to 128 bytes in size. It achieves this by tracking size
information, using an extra line to do so. Only one bit per word is
necessary to track size information, because we only care if an object
crosses 1 line boundary or 0.

This reduces the overhead of the immix simulation from about 8.5% to
7.0%.